refactor: clean DocumentationTable
